Move allocateTimer from base to core
